Attempting data connection
