+++ mysql.properties
@@ -1,4 +1,4 @@
-//jdbc:mysql://localhost:3306/sakila", "root", "password"
-db.url=jdbc:mysql://localhost:3306/sakila
+
+db.url=jdbc:mysql://localhost:3306/employeelist
 db.userID=root
 db.password=password
